move message builder to resources folder
